Split widget: computed detail
